pyproject.toml
[tool.poetry]
name = "yanashbelyak"
version = "0.1.0"
description = ""
authors = ["Alexander Rakhmaev <rahmaevao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
pelican = "^4.9.1"
singleton-decorator = "^1.0.0"
markdown = "^3.5.1"
drawsvg = "^2.3.0"
ghp-import = "^2.1.0"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.9"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
